% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgscore.R
\name{avgscore}
\alias{avgscore}
\title{avgscore}
\usage{
avgscore(data)
}
\arguments{
\item{data}{data frame of anime stats created with loop()}
}
\value{
tells you the average rating of animes in your data frame
}
\description{
avgscore
}
