% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createanimexl.R
\name{createanimexl}
\alias{createanimexl}
\title{createanimexl}
\usage{
createanimexl(start, end, reference = "anime", request = "stats")
}
\arguments{
\item{start}{id of the first anime you would like stats on}

\item{end}{id of the last anime you would like stats on}

\item{request}{}
}
\value{
creates an excel spreadsheet of the data frame created with loop()
}
\description{
createanimexl
}
