% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isthereonepiece.R
\name{isthereonepiece}
\alias{isthereonepiece}
\title{isthereonepiece}
\usage{
isthereonepiece(data)
}
\arguments{
\item{data}{data frame created with loop function of anime data}
}
\value{
returns true of One Piece is one of the anime in the data frame, false if it is not
}
\description{
isthereonepiece
}
