% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jikan_get.R
\name{jikan_get}
\alias{jikan_get}
\title{jikin_get}
\usage{
jikan_get(reference = "anime", id = "1", request = "", parameter = "")
}
\arguments{
\item{id}{id number of the anime series you would like data on}

\item{parameter}{}
}
\value{

}
\description{
jikin_get
}
