% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.R
\name{loop}
\alias{loop}
\title{loops}
\usage{
loop(start, end, reference = "anime", request = "stats")
}
\arguments{
\item{start}{id of the first anime you want stats on}

\item{end}{id of the last anime you want stats on}

\item{request}{}
}
\value{
returns data frame of stats on the selected anime
}
\description{
loops
}
